VTXdigi: chasing (&fixing) an error where the cluster pos would lie outside of the sensor because of floating point noise. I suspect it was the back-and-forth between float and double in the path computation, the errors didnt show up after moving the whole path computation to double. Added a less eager throw and clamping on top.

+++ VTXdigi_Modular/src/ChargeCollector_impl.h
@@ -33,7 +33,7 @@ struct Path {
 };
 
 /** @brief Compute the factors by which to clip a path along a given axis (to clip it to the sensor volume) */
-std::array<float, 2> ComputePathClippingFactors(std::array<float, 2> t, const float entry_ax, const float travel_ax, const float sensorLength_ax);
+std::array<double, 2> ComputePathClippingFactors(std::array<double, 2> t, const double entry_ax, const double travel_ax, const double sensorLength_ax);
 
 /** @brief Construct path information from a simHit and the sensor's transformation matrix
  * @note returns true if path is valid, false otherwise. false means the path would not intersect the sensor volume */
